requirements.txt
streamlit
googlemaps
folium
streamlit-folium
opencv-python
pyaudio
wave
twilio
geopy
openai
pillow
numpy
